+++ CHANGELOG.md
@@ -1,8 +1,15 @@
-# Release Notes for 1.x
+# Release Notes for 2.x
+
+## v2.0.1 2025-08-01
+
+* Fix grant entities through entities granted by default by [@braxey](https://github.com/braxey) in https://github.com/gillyware/gatekeeper/commit/6510e403957cf71ecbb2f00600d5516c5ff2c41d
+* Fix team command by [@braxey](https://github.com/braxey) in https://github.com/gillyware/gatekeeper/commit/d08e5beac9c532314c6a4b9f8dd6c0785ebbbd2a
 
 ## v2.0.0 - 2025-08-01
 
 * Implement denials and granting by default for all entities by [@braxey](https://github.com/braxey) in https://github.com/gillyware/gatekeeper/commit/7d960875f899829f423336537a98275b11c31430
+
+# Release Notes for 1.x
 
 ## v1.1.0 - 2025-07-27